% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{ahp}
\alias{ahp}
\title{AHP ranking using eigenvalue based method}
\usage{
ahp(M, ...)
}
\arguments{
\item{M}{- n x n criteria matrix}

\item{...}{- matrixes with the comparisons of alternatives with respect to the given criteria}
}
\value{
multicriteria ranking using eigenvalue based method
}
\description{
Computes milticriteria ranking using eigenvalue based method using criteria matrix and alternatives matrixes.
This is basic three levels AHP.
}

