% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{AIJadd}
\alias{AIJadd}
\title{Aggregation matrix/vector (aritmetic means)}
\usage{
AIJadd(...)
}
\arguments{
\item{...}{- matrices od vectors the same dimension}
}
\value{
aggregation matrix/vector
}
\description{
Computes aggregation matrix/vector whose elements are aritmethic means
of elements from input matrices/vectors (whose need to be the same dimension)
}

