% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{AIJgeom}
\alias{AIJgeom}
\title{Aggregation matrix/vector (geometric means)}
\usage{
AIJgeom(...)
}
\arguments{
\item{...}{- matrices od vectors the same dimension}
}
\value{
aggregation matrix/vector
}
\description{
COmputes aggregation matrix/vector whose elements are aritmethic means
of elements from input matrices/vectors (whose need to be the same dimension)
}

