% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{consistentMatrixFromRank}
\alias{consistentMatrixFromRank}
\title{Consistent matrix from rank}
\usage{
consistentMatrixFromRank(rankList)
}
\arguments{
\item{rankList}{- ranking of matrix}
}
\value{
- full PC matrix
}
\description{
Creates consistent PC matrix from the rank vector V so that m_ij = v[i]/v[j]
}

