% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{cop1Check}
\alias{cop1Check}
\title{check COP1}
\usage{
cop1Check(matrix, resultList)
}
\arguments{
\item{matrix}{- PC matrix}

\item{resultList}{- ranking of matrix}
}
\value{
true if the first Condition of Order Preservation  is fulfilled, else false
}
\description{
checks if the first Condition of Order Preservation  is fulfilled
}

