% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{cop1ViolationList}
\alias{cop1ViolationList}
\title{List of COP1}
\usage{
cop1ViolationList(matrix, resultList)
}
\arguments{
\item{matrix}{- PC matrix}

\item{resultList}{- ranking of matrix}
}
\value{
the list of indices that violate the first Condition of Order Preservation
}
\description{
counts the list of indices that violate the first Condition of Order Preservation
postulate (formulated by Bana e Costa and Vansnick)
}

