% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{cop2Check}
\alias{cop2Check}
\title{check COP2}
\usage{
cop2Check(matrix, resultList)
}
\arguments{
\item{matrix}{- PC matrix}

\item{resultList}{- ranking of matrix}
}
\value{
true if the second Condition of Order Preservation is fulfilled, else false
}
\description{
checks if the second Condition of Order Preservation  is fulfilled
}

