% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{cop2ViolationList}
\alias{cop2ViolationList}
\title{List of COP2}
\usage{
cop2ViolationList(matrix, resultList)
}
\arguments{
\item{matrix}{- PC matrix}

\item{resultList}{- ranking of matrix}
}
\value{
the list of indices that violate the second Condition of Order Preservation
}
\description{
counts the list of indices that violate the second Condition of Order Preservation
postulate (formulated by Bana e Costa and Vansnick)
}

