% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{deleteColumns}
\alias{deleteColumns}
\title{Delete columns}
\usage{
deleteColumns(matrix, listOfColumns)
}
\arguments{
\item{listOfColumns}{- indices of columns which need to be deleted}

\item{martix}{- matrix}
}
\value{
matrix without deleted columns
}
\description{
Delete columns from matrix
}

