% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{deleteRows}
\alias{deleteRows}
\title{Delete rows}
\usage{
deleteRows(matrix, listOfRows)
}
\arguments{
\item{listOfRows}{- indices of rows which need to be deleted}

\item{martix}{- PC matrix}
}
\value{
matrix without deleted rows
}
\description{
Delete rows from matrix
}

