% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{deleteRowsAndColumns}
\alias{deleteRowsAndColumns}
\title{Delete rows and columns}
\usage{
deleteRowsAndColumns(matrix, listOfRowsAndColumns)
}
\arguments{
\item{listOfRowsAndColumns}{- indices of rows columns which need to be deleted}

\item{martix}{- matrix}
}
\value{
matrix without deleted rows and columns
}
\description{
Delete columns and rows from matrix
}

