% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{eigenValueRank}
\alias{eigenValueRank}
\title{Rescaled principal eigen vector of matrix}
\usage{
eigenValueRank(matrix)
}
\arguments{
\item{matrix}{- matrix}
}
\value{
the rescaled principal eigen vector of matrix
}
\description{
Returns the principal eigen vector of matrix rescaled in a way that the sum of its entries is 1
}

