% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{eigenValueRankSym}
\alias{eigenValueRankSym}
\title{A symbolic version of function eigenValueRank}
\usage{
eigenValueRankSym(matrix)
}
\arguments{
\item{matrix}{- matrix}
}
\value{
the principal eigen vector of matrix rescaled in a way that the sum of its entries is 1
}
\description{
A symbolic version of function eigenValueRank
}

