% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{errorMatrix}
\alias{errorMatrix}
\title{Ranking discrepancy}
\usage{
errorMatrix(matrix, mju)
}
\arguments{
\item{mju}{- ranking of matrix}

\item{M}{- matrix}
}
\value{
error matrix E = [e_ij]
}
\description{
Computes discrepancy for each element of matrix
and create matrix containing entries in form e_ij = m_ji*(mju_i/mju_j).
When the M matrix is consistent every r_ij equals 1.
}

