% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{geometricRank}
\alias{geometricRank}
\title{Rank list given as geometric means}
\usage{
geometricRank(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
The rank list given as geometric means
}
\description{
Returns rank list given as geometric means of rows of the matrix
}

