% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{geometricRescaledRank}
\alias{geometricRescaledRank}
\title{Rescaled rank list given as geometric means}
\usage{
geometricRescaledRank(matrix)
}
\arguments{
\item{matrix}{- matrix}
}
\value{
Rescaled rank list given as geometric means of rows of the matrix
}
\description{
Returns rank list given as geometric means of rows of the matrix rescaled in way
that the sum of its entries is 1
}

