% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{getMatrixEntry}
\alias{getMatrixEntry}
\title{Element of matrix}
\usage{
getMatrixEntry(matrix, r, c)
}
\arguments{
\item{matrix}{- matrix}

\item{r}{- row}

\item{c}{- column}
}
\value{
[r,c] element of matrix
}
\description{
Returns [r,c] element of matrix
}

