% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{globalDiscrepancy}
\alias{globalDiscrepancy}
\title{Global discrepancy}
\usage{
globalDiscrepancy(matrix, mju)
}
\arguments{
\item{matrix}{- matrix}

\item{mju}{- ranking of matrix}
}
\value{
maximal value of local discrepancy
}
\description{
Finds maximal entry of local discrepancy matrix
}

