% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{harkerMatrix}
\alias{harkerMatrix}
\title{Harher matrix}
\usage{
harkerMatrix(matrix)
}
\arguments{
\item{matrix}{- matrix}
}
\value{
matrix which is ready to use in eigenvalue based method
}
\description{
Computes harker matrix A for a given incomplete PC matrix.
Then matrix A is ready to use in eigenvalue based method.
Empty and incorrect values are replaced by 0 and diagonal values are replaced by 1.
}

