% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{harkerMatrixPlaceHolderCount}
\alias{harkerMatrixPlaceHolderCount}
\title{Number of zeros elements in row}
\usage{
harkerMatrixPlaceHolderCount(matrix, row)
}
\arguments{
\item{matrix}{- matrix}

\item{row}{- number of row which will be checked}
}
\description{
Counts number of zeros elements in given row in matrix
}

