% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{HREgeomConstantTermVector}
\alias{HREgeomConstantTermVector}
\title{constant term vector for the HRE geometric method}
\usage{
HREgeomConstantTermVector(matrix, knownVector)
}
\arguments{
\item{matrix}{- PC matrix}

\item{knownVector}{- vector of known alternatives and others are marked by value 0}
}
\value{
constant term vector b for the HRE method
}
\description{
HRE constant term vector together with HRE matrix forms the linear equation system Au=b
}

