% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{HREgeomFullRank}
\alias{HREgeomFullRank}
\title{full rank HRE using geometric method}
\usage{
HREgeomFullRank(matrix, knownVector)
}
\arguments{
\item{matrix}{- PC matrix}

\item{knownVector}{- vector of known alternatives and others are marked by value 0}
}
\value{
values for the both: known and unknown alternatives
}
\description{
Counts values for the unknown alternatives and add their to known alternatives
}

