% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{HREgeomMatrix}
\alias{HREgeomMatrix}
\title{Matrix for the HRE geometric method}
\usage{
HREgeomMatrix(matrix, knownVector)
}
\arguments{
\item{matrix}{- PC matrix}

\item{knownVector}{- vector of known alternatives and others are marked by value 0}
}
\value{
Matrix A for the HRE method
}
\description{
HRE matrix together with HRE constant term vector forms the linear equation system Au=b
}

