% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{HREgeomRescaledRank}
\alias{HREgeomRescaledRank}
\title{Rescaled HRE full rank using geometric method}
\usage{
HREgeomRescaledRank(matrix, knownVector)
}
\arguments{
\item{matrix}{- PC matrix}

\item{knownVector}{- vector of known alternatives and others are marked by value 0}
}
\value{
the rescaled HRE full rank list
}
\description{
Counts the rescaled HRE full rank list. Rescaled in the entries sum up to 1
}

