% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{HREpartialRank}
\alias{HREpartialRank}
\title{Unknown HRE values}
\usage{
HREpartialRank(matrix, knownVector)
}
\arguments{
\item{matrix}{- PC matrix}

\item{knownVector}{- vector of known alternatives and others are marked by value 0}
}
\value{
values for the unknown alternatives
}
\description{
Counts values for the unknown alternatives
}

