% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{koczkodajConsistentTriad}
\alias{koczkodajConsistentTriad}
\title{Consistent triad}
\usage{
koczkodajConsistentTriad(triad)
}
\arguments{
\item{triad}{- vector of three values}
}
\value{
the closest consistenr triad
}
\description{
Counts the closest consistenr triad
}

