% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{koczkodajIdx}
\alias{koczkodajIdx}
\title{Koczkodaj inconsistency}
\usage{
koczkodajIdx(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
Koczkodaj inconsistency
}
\description{
Counts the value of Koczkodaj inconsistency for the matrix
}

