% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{koczkodajImprovedMatrixStep}
\alias{koczkodajImprovedMatrixStep}
\title{More consistent matrix}
\usage{
koczkodajImprovedMatrixStep(matrix)
}
\arguments{
\item{matric}{- PC matrix}
}
\value{
more consistent matrix
}
\description{
Counts improved (more consistent) matrix in which the most inconsistent triad is replaced by tje closest consistent one
}

