% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{koczkodajTheWorstTriad}
\alias{koczkodajTheWorstTriad}
\title{The most inconsistent triad}
\usage{
koczkodajTheWorstTriad(matrix)
}
\arguments{
\item{matrix}{- PC Matrix}
}
\value{
the worst triad in matrix according to Koczkodaj inconsistency criterion
}
\description{
Counts the worst triad in matrix according to Koczkodaj inconsistency criterion
}

