% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{koczkodajTheWorstTriads}
\alias{koczkodajTheWorstTriads}
\title{The most inconsistent triads}
\usage{
koczkodajTheWorstTriads(matrix, n)
}
\arguments{
\item{matrix}{- PC Matrix}
}
\value{
n worst triads in matrix according to Koczkodaj inconsistency criterion
}
\description{
Counts the worst triads in matrix according to Koczkodaj inconsistency criterion
}

