% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{koczkodajTriadIdx}
\alias{koczkodajTriadIdx}
\title{Koczkodaj triad innconsistency}
\usage{
koczkodajTriadIdx(triad)
}
\arguments{
\item{triad}{- vector of 3 elements}
}
\value{
the Koczkodaj triad inconsistency
}
\description{
Returns the Koczkodaj triad inconsistency
}

