% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{localDiscrepancyMatrix}
\alias{localDiscrepancyMatrix}
\title{Local discrepancy}
\usage{
localDiscrepancyMatrix(matrix, mju)
}
\arguments{
\item{matrix}{- matrix}

\item{mju}{- ranking of matrix}
}
\value{
matrix with locals discrepancy
}
\description{
Compute matrix with entries d_ij = max{e_ij-1, 1/e_ij-1}
}

