% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{normalizedKendallTauDistance}
\alias{normalizedKendallTauDistance}
\title{Normalized Kendall Tau distance for two vectors}
\usage{
normalizedKendallTauDistance(list1, list2)
}
\arguments{
\item{list1}{- first rank to compare}

\item{list2}{- second rank to compare}
}
\value{
proportion of Kendall Tau distance to numbers of all possible pairs
}
\description{
Computes Kendall (bubble sort) distance between two rank vectors and divited it by numbers of all possible pairs
}

