% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{principalEigenValue}
\alias{principalEigenValue}
\title{The largest principal eigenvalue}
\usage{
principalEigenValue(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
the largest principal eigenvalue of matrix
}
\description{
Returns the largest principal eigenvalue of matrix
}

