% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{principalEigenVector}
\alias{principalEigenVector}
\title{Eigenvector of matrix}
\usage{
principalEigenVector(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
the eigenvector of matrix corresponding to itd principal eigenvaluex
}
\description{
Returns the eigenvector of matrix corresponding to itd principal eigenvalue
}

