% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{principalEigenVectorSym}
\alias{principalEigenVectorSym}
\title{Eigenvector of matrix (symbolic version)}
\usage{
principalEigenVectorSym(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
the eigenvector of matrix corresponding to itd principal eigenvaluex
}
\description{
A symbolic version of function principalEigenVector
}

