% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{recreatePCMatrix}
\alias{recreatePCMatrix}
\title{Recipropal matrix}
\usage{
recreatePCMatrix(matrix)
}
\arguments{
\item{matrix}{- matrix}
}
\value{
recipropal matrix
}
\description{
Recreates recipropal matrix of the basis of upper-triagle of matrix
}

