% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{saatyIdx}
\alias{saatyIdx}
\title{Value of the Saaty Inconsistency Index}
\usage{
saatyIdx(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
the value of the Saaty Inconsistency Index computed for the matrix
}
\description{
Returns the value of the Saaty Inconsistency Index computed for the matrix
}

