% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{saatyIdxSym}
\alias{saatyIdxSym}
\title{A symbolic version of function saatyIdxSym}
\usage{
saatyIdxSym(matrix)
}
\arguments{
\item{matrix}{- matrix}
}
\value{
the value of the Saaty Inconsistency Index computed for the matrix
}
\description{
A symbolic version of function saatyIdxSym
}

