% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{setDiagonal}
\alias{setDiagonal}
\title{Set diagonal of matrix}
\usage{
setDiagonal(matrix, value)
}
\arguments{
\item{matrix}{- matrix}

\item{value}{- value which need to be set on the diagonal}
}
\value{
Matrix with the specifed value on the diagonal
}
\description{
Sets diagonal of matrix to specifed value
}

