% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{principalEigenValueSym}
\alias{principalEigenValueSym}
\title{The largest principal eigenvalue (symbolic version)}
\usage{
principalEigenValueSym(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
the largest principal eigenvalue of matrix
}
\description{
A symbolic version of function principalEigenValue
}

