% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{principalEigenVectorSym}
\alias{principalEigenVectorSym}
\title{A symbolic version of function principalEigenVector}
\usage{
principalEigenVectorSym(matrix)
}
\arguments{
\item{matrix}{- matrix}
}
\value{
the eigenvector of matrix corresponding to itd principal eigenvaluex
}
\description{
A symbolic version of function principalEigenVector
}

