% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{saatyIdxSym}
\alias{saatyIdxSym}
\title{Value of the Saaty Inconsistency Index (symbolic version)}
\usage{
saatyIdxSym(matrix)
}
\arguments{
\item{matrix}{- PC matrix}
}
\value{
the value of the Saaty Inconsistency Index computed for the matrix
}
\description{
A symbolic version of function saatyIdxSym
}

